// API_Client/client/models/product-color-item-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ProductColorItemDto {
    /**
     * Color-ის ID (წინასწარ /colors-ზე შექმნილი)
     */
    'colorId': string;
    /**
     * ამ ფერის მარაგი საწყობში (ცალი)
     */
    'stock': number;
}